requirements.txt
astronomer-cosmos[dbt]